workers/requirements.txt
redis>=5.0.0
aioredis>=2.0.0
psycopg2-binary>=2.9.0
asyncio>=3.4.3
# LlamaCloud for document parsing (EU: api.cloud.eu.llamaindex.ai)
llama-cloud>=0.1.0
llama-parse>=0.4.0
# Moonshot AI (uses OpenAI-compatible API)
openai>=1.0.0
pageindex>=0.2.5
tenacity>=8.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
# Adverse Media Check - Free web scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
